% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertImage.R
\name{insertImage}
\alias{insertImage}
\title{Insert image.}
\usage{
insertImage()
}
\value{
Inserts \code{![alt](path/to/file)}. Optionally copies file to project.
}
\description{
Addin for RStudio for inserting an image into Markdown.

Copy the image file to the project in img/ or /vignette directory.

Uses relative paths for image files in project and absolute paths for images not in project.

See \code{details} for setting up key command.
}
\details{
How to set up key command in RStudio:

After installing package.
Go to:

\code{Tools} >> \code{Addins} >> \code{Browse Addins} >> \code{Keyboard Shortcuts}.

Find \code{"Insert Image"} and press its field under \code{Shortcut}.

Press desired key command.

Press \code{Apply}.

Press \code{Execute}.
}
\author{
Ludvig Renbo Olsen, \email{r-pkgs@ludvigolsen.dk}
}

